# docs/swagger.yaml
definitions:
  ActivationByCode.ActivationByCodeCommand:
    properties:
      activationCode:
        type: integer
    type: object
  Common.ResponseModel-GetProductDetail_GetProductQueryVM:
    properties:
      data:
        $ref: '#/definitions/GetProductDetail.GetProductQueryVM'
      message:
        type: string
      statusCode:
        type: integer
    type: object
  Common.ResponseModel-array_GetCategories_GetCategoriesVm:
    properties:
      data:
        items:
          $ref: '#/definitions/GetCategories.GetCategoriesVm'
        type: array
      message:
        type: string
      statusCode:
        type: integer
    type: object
  Common.ResponseModel-string:
    properties:
      data:
        type: string
      message:
        type: string
      statusCode:
        type: integer
    type: object
  ForgetPassword.ForgetPasswordCommand:
    properties:
      phoneNumber:
        type: integer
    type: object
  GetCategories.GetCategoriesVm:
    properties:
      id:
        type: integer
      likeCount:
        type: integer
      parentCatId:
        type: integer
      summary:
        type: string
      title:
        type: string
    type: object
  GetCategoryProducts.GetCategoryProductsQuery:
    properties:
      categoryId:
        type: integer
      count:
        type: integer
      sortType:
        type: string
      start:
        type: integer
    type: object
  GetProductDetail.GetProductQueryVM:
    properties:
      categoryId:
        type: integer
      existCount:
        type: integer
      id:
        type: integer
      likeCount:
        type: integer
      maxPrice:
        type: integer
      minPrice:
        type: integer
      properties:
        items:
          $ref: '#/definitions/GetProductDetail.ProductPropertiesVM'
        type: array
      rankNumber:
        type: number
      selectables:
        $ref: '#/definitions/GetProductDetail.ProductSelectablesVM'
      summary:
        type: string
      title:
        type: string
    type: object
  GetProductDetail.ProductPropertiesVM:
    properties:
      propertyValue:
        type: string
      title:
        type: string
    type: object
  GetProductDetail.ProductSelectablesVM:
    properties:
      title:
        type: string
      values:
        items:
          $ref: '#/definitions/GetProductDetail.ProductSelectablesValuesVM'
        type: array
    type: object
  GetProductDetail.ProductSelectablesValuesVM:
    properties:
      existCount:
        type: integer
      valueKey:
        type: string
      valueTitle:
        type: string
    type: object
  Register.RegisterCommand:
    properties:
      Password:
        type: string
      PhoneNumber:
        type: integer
    type: object
info:
  contact: {}
  version: 1.0.0// @host localhost:8080
paths:
  /Account/ActivateUser:
    post:
      parameters:
      - description: ActivationData
        in: body
        name: ActivateUser
        required: true
        schema:
          $ref: '#/definitions/ActivationByCode.ActivationByCodeCommand'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Common.ResponseModel-string'
      security:
      - bearer: []
      summary: Activate User by Code
      tags:
      - Account
  /Account/ForgetPassword:
    post:
      parameters:
      - description: ForgetPassword
        in: body
        name: ForgetPassword
        required: true
        schema:
          $ref: '#/definitions/ForgetPassword.ForgetPasswordCommand'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Common.ResponseModel-string'
      security:
      - bearer: []
      summary: Register an User
      tags:
      - Account
  /Account/Register:
    post:
      parameters:
      - description: RegisterData
        in: body
        name: register
        required: true
        schema:
          $ref: '#/definitions/Register.RegisterCommand'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Common.ResponseModel-string'
      security:
      - bearer: []
      summary: Register an User
      tags:
      - Account
  /Categories/GetCategories:
    get:
      parameters:
      - description: Parent Id
        in: query
        name: parentId
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Common.ResponseModel-array_GetCategories_GetCategoriesVm'
      security:
      - bearer: []
      summary: Get Categories List
      tags:
      - Categories
  /Categories/GetCategoryProducts:
    post:
      parameters:
      - description: GetCategoryProduct
        in: body
        name: GetCategoryProduct
        required: true
        schema:
          $ref: '#/definitions/GetCategoryProducts.GetCategoryProductsQuery'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Common.ResponseModel-array_GetCategories_GetCategoriesVm'
      security:
      - bearer: []
      summary: Get Category Products
      tags:
      - Categories
  /Products/GetProductDetail:
    get:
      parameters:
      - description: productId
        in: query
        name: productId
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Common.ResponseModel-GetProductDetail_GetProductQueryVM'
      security:
      - bearer: []
      summary: Get Product Detail
      tags:
      - Products
securityDefinitions:
  bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"